Move TrackSetTableModel into trackset folder
