+++ src/library/trackset/crate/cratetablemodel.h
@@ -2,7 +2,7 @@
 
 #include "library/basesqltablemodel.h"
 #include "library/trackset/crate/crateid.h"
-#include "library/tracksettablemodel.h"
+#include "library/trackset/tracksettablemodel.h"
 
 class CrateTableModel : public TrackSetTableModel {
     Q_OBJECT
